fixed inconsistency
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@26094 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/traci/TraCIAPI.h
@@ -176,7 +176,7 @@ public:
         std::vector<std::string> getLastStepVehicleIDs(const std::string& edgeID) const;
 
         void adaptTraveltime(const std::string& edgeID, double time, int beginSeconds = 0, int endSeconds = std::numeric_limits<int>::max()) const;
-        void setEffort(const std::string& edgeID, double effort, int beginSeconds = 0, int endSeconds = std::numeric_limits<double>::max()) const;
+        void setEffort(const std::string& edgeID, double effort, int beginSeconds = 0, int endSeconds = std::numeric_limits<int>::max()) const;
         void setMaxSpeed(const std::string& edgeID, double speed) const;
 
     private:
